+++ src/sodium.rs
@@ -25,7 +25,7 @@ use exonum_sodiumoxide::crypto::{
 };
 use serde::{Deserialize, Serialize};
 
-use super::{Cipher, CipherOutput, DeriveKey, Eraser, Suite};
+use crate::{Cipher, CipherOutput, DeriveKey, Eraser, MacMismatch, Suite};
 
 /// `Scrypt` key derivation function parameterized as per libsodium, i.e., via
 /// `opslimit` (computational hardness) and `memlimit` (RAM consumption).
@@ -143,13 +143,18 @@ impl Cipher for XSalsa20Poly1305 {
         }
     }
 
-    fn open(output: &mut [u8], enc: &CipherOutput, nonce: &[u8], key: &[u8]) -> Result<(), ()> {
+    fn open(
+        output: &mut [u8],
+        enc: &CipherOutput,
+        nonce: &[u8],
+        key: &[u8],
+    ) -> Result<(), MacMismatch> {
         let nonce = Nonce::from_slice(nonce).expect("invalid nonce length");
         let key = Key::from_slice(key).expect("invalid key length");
         let mac = Tag::from_slice(&enc.mac).expect("invalid MAC length");
 
         output.copy_from_slice(&enc.ciphertext);
-        open_detached(output, &mac, &nonce, &key)
+        open_detached(output, &mac, &nonce, &key).map_err(|()| MacMismatch)
     }
 }
 
@@ -174,13 +179,18 @@ impl Cipher for ChaCha20Poly1305 {
         }
     }
 
-    fn open(output: &mut [u8], enc: &CipherOutput, nonce: &[u8], key: &[u8]) -> Result<(), ()> {
+    fn open(
+        output: &mut [u8],
+        enc: &CipherOutput,
+        nonce: &[u8],
+        key: &[u8],
+    ) -> Result<(), MacMismatch> {
         let nonce = aead::Nonce::from_slice(nonce).expect("invalid nonce length");
         let key = aead::Key::from_slice(key).expect("invalid key length");
         let mac = aead::Tag::from_slice(&enc.mac).expect("invalid MAC length");
 
         output.copy_from_slice(&enc.ciphertext);
-        aead::open_detached(output, None, &mac, &nonce, &key)
+        aead::open_detached(output, None, &mac, &nonce, &key).map_err(|()| MacMismatch)
     }
 }