+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 ### Changed
 
 - Make non-instantiated types (such as cryptosuites) to structs.
+- Use a separate error type, `MacMismatch`, instead of `()` in `Cipher::open()`.
 - Update dependencies.
 
 ## 0.3.0 - 2020-03-11